pyproject.toml
[tool.poetry]
name = "strawberry-django-jwt"
version = "0.1.1-a3"
description = "Strawberry-graphql port of the graphene-django-jwt package"
license = "MIT"
authors = ["KundaPanda <vojdoh@gmail.com>"]
maintainers = []
readme = "README.md"
classifiers = [
    'Development Status :: 3 - Alpha',
    'Environment :: Web Environment',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Framework :: Django',
    'Framework :: Django :: 3.0',
    'Framework :: Django :: 3.1',
    'Framework :: Django :: 3.2',
]
exclude = ["tests/"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.0"
PyJWT = ">=1.7.1,<3.0"
strawberry-graphql = ">=0.65.0,<0.70.0"
strawberry-graphql-django = ">=0.2.0,<0.3.0"
django-admin-display = "^1.3.0"
packaging = "^21.0"
importlib-metadata = {version = "^4.0.0", python = "<=3.7"}
[tool.poetry.dev-dependencies]
cryptography = "^3.0"
coverage = "^5.0"
pytest = "^6.0"
pytest-cov = "^2.0"
pytest-django = "^4.0.0"
nox = "^2021.6.12"
types-cryptography = "^3.3.3"
django-stubs = "^1.8.0"
django-mock-queries = "^2.1.6"
types-mock = "^0.1.3"
types-jwt = "^0.1.3"
types-pkg-resources = "^0.1.3"
pre-commit = "^2.13.0"
pre-commit-hooks = "^4.0.1"
autopep8 = "^1.4"
darglint = "^1.8.0"
mypy = "^0.910"
safety = "^1.5"
pylint = "^2.9.3"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = "tests.py test_*.py *_tests.py"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
